Placeholder HUD created, Weapon first steps and network first steps
Placeholders of prematch menu widgets and ingame HUD. Weapon prototyping (not yet functional and in the absence of refactoring hierarchy to be able to have easily more weapons having an interface)

Start of first actions necessary to mount the game in network (UMG has already been prepared in gamemode). Marked Character for non-replication.